// showcases/stock-pulse/src/api.types.ts
// Auto-generated by Vitek - DO NOT EDIT
// This file is automatically generated from your API routes

export type VitekParams = Record<string, string>;
export type VitekQuery = Record<string, string | string[]>;


export type MovementsPostBody = {
  sku: string;
  kind: 'in' | 'out';
  quantity: number;
};

export interface HealthGetParams extends VitekParams {

}

export interface MovementsPostParams extends VitekParams {

}

export interface StockSkuGetParams extends VitekParams {
  sku: string;
}

export interface StockGetParams extends VitekParams {

}

export type VitekRoute =
  | {
      pattern: 'health';
      method: 'get';
      params: HealthGetParams;
    }
  | {
      pattern: 'movements';
      method: 'post';
      params: MovementsPostParams;
    }
  | {
      pattern: 'stock/:sku';
      method: 'get';
      params: StockSkuGetParams;
    }
  | {
      pattern: 'stock';
      method: 'get';
      params: StockGetParams;
    };

export const API_BASE_PATH = '/api' as const;